ALOT2

+++ ASPForum/Controllers/ReplyController.cs
@@ -19,16 +19,60 @@ namespace ASPForum.Controllers
             return View();
         }
 
+        [HttpGet]
+        public ActionResult Edit(int categoryId, int threadId, int replyId)
+        {
+            Thread thread = db.Threads.Include("Subject").FirstOrDefault(thrd => thrd.Id == threadId);
+            Reply reply = db.Replies.Include("Author").FirstOrDefault(repl => repl.Id == replyId);
+            ViewBag.CategoryId = categoryId;
+            ViewBag.ThreadId = thread.Id;
+            ViewBag.SubjectId = thread.Subject.Id;
+            ViewBag.Reply = reply.Id;
+            return View(reply);
+        }
+
+
+        [HttpPut]
+        public ActionResult Edit(int categoryId, int subjectId, int threadId, Reply newReply)
+        {
+            try
+            {
+                Thread thread = db.Threads.Find(threadId);
+                Reply reply = db.Replies.Include("Author").Where(repl => repl.Id == newReply.Id).FirstOrDefault();
+                if (reply.Author.Id == User.Identity.GetUserId())
+                {
+                    reply.Content = newReply.Content;
+                    db.SaveChanges();
+                }
+                else
+                {
+                    return HttpNotFound();
+                }
+                return Redirect("/Thread/Show/" + categoryId + "/" + subjectId + "/" + thread.Id);
+            }
+            catch (Exception e)
+            {
+                return View();
+            }
+        }
+
 
         [HttpDelete]
-        [Authorize(Roles = "Administrator, Moderator")]
-        public ActionResult Delete(int id) {
-            Reply reply = db.Replies.Find(id);
-            if (reply.Author.Id == User.Identity.GetUserId() || User.IsInRole("Administrator") || User.IsInRole("Moderator")) {
+        [Authorize(Roles = "User, Administrator, Moderator")]
+        public ActionResult Delete(int categoryId, int subjectId, int replyId)
+        {
+            Reply reply = db.Replies.Include("Author").Include("Thread").FirstOrDefault(repl => repl.Id == replyId);
+            if ((reply.Author != null && reply.Author.Id == User.Identity.GetUserId()) || User.IsInRole("Administrator") || User.IsInRole("Moderator"))
+            {
+                var threadId = reply.Thread.Id;
                 db.Replies.Remove(reply);
                 db.SaveChanges();
+                return Redirect("/Thread/Show/" + categoryId + "/" + subjectId + "/" + threadId);
             }
-            return Redirect("/Category/Index");
+            else
+            {
+                return HttpNotFound();
+            }
         }
